refactor: add Configurable trait
