+++ src/common/config/Cargo.toml
@@ -9,6 +9,22 @@ workspace = true
 
 [dependencies]
 common-base.workspace = true
+common-error.workspace = true
+common-macro.workspace = true
+config.workspace = true
 num_cpus.workspace = true
 serde.workspace = true
 sysinfo.workspace = true
+snafu.workspace = true
+serde_json.workspace = true
+toml.workspace = true
+
+[dev-dependencies]
+common-test-util.workspace = true
+serde.workspace = true
+common-wal.workspace = true
+common-telemetry.workspace = true
+datanode.workspace = true
+tempfile.workspace = true
+meta-client.workspace = true
+temp-env = "0.3"
