+++ src/datanode/Cargo.toml
@@ -18,6 +18,7 @@ bytes.workspace = true
 catalog.workspace = true
 client.workspace = true
 common-base.workspace = true
+common-config.workspace = true
 common-error.workspace = true
 common-function.workspace = true
 common-greptimedb-telemetry.workspace = true
