+++ src/frontend/src/error.rs
@@ -343,8 +343,9 @@ pub enum Error {
 
     #[snafu(display("Failed to serialize options to TOML"))]
     TomlFormat {
-        #[snafu(source)]
-        error: toml::ser::Error,
+        #[snafu(implicit)]
+        location: Location,
+        source: common_config::error::Error,
     },
 
     #[snafu(display("Failed to get cache from cache registry: {}", name))]
