+++ src/meta-srv/Cargo.toml
@@ -16,6 +16,7 @@ async-trait = "0.1"
 client.workspace = true
 common-base.workspace = true
 common-catalog.workspace = true
+common-config.workspace = true
 common-error.workspace = true
 common-greptimedb-telemetry.workspace = true
 common-grpc.workspace = true
